feat: like 게시글에 종속된 형태로 변경

+++ src/main/java/com/kakaotech/ott/ott/like/domain/repository/LikeRepository.java
@@ -13,7 +13,7 @@ public interface LikeRepository {
 
     boolean existsByUserIdAndPostId(Long userId, Long postId);
 
-    int findByPostId(Long postId);
+    Long findByPostId(Long postId);
 
     Set<Long> findLikedPostIdsByUserId(Long userId, List<Long> postIds);
 }
